refactor: extract TEE token client into shared tee_token module
